MAIN-32888 - update schema when changing mapping
AddTableWithNameAndSchema supports being called for structs that are already in
the table map. When that happens, it updates the mapping in place. While it was
correctly updating the table name, it wasn't updating the schema name, and also
wasn't clearing out the cached queries. Both of those things need to happen for
a clean update of an existing mapping.